auth cookie logic moved to auth.ts
